requirements.txt
torch>=2.2
torchvision>=0.17
datasets>=2.19
clearml>=1.16
numpy==2.2.6
scikit-learn>=1.4
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.66
Pillow>=10.0
gradio==5.49.1
huggingface_hub<1.1.4
black
isort
pre-commit
pytest
